adding data validation to deletes

+++ src/main/java/com/apex/swimtime/service/SwimTimeService.java
@@ -1,17 +1,17 @@
 package com.apex.swimtime.service;
 
-import com.apex.swimtime.constants.Split;
-import com.apex.swimtime.constants.StrokeEnum;
-import com.apex.swimtime.constants.SwimTime;
-import com.apex.swimtime.constants.SwimTimeRO;
+import com.apex.swimtime.constants.*;
 import com.apex.swimtime.repository.SplitRepository;
 import com.apex.swimtime.repository.SwimTimeRepository;
 import com.apex.swimtime.repository.SwimmerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class SwimTimeService {
@@ -125,31 +125,45 @@ public class SwimTimeService {
     }
 
     public Integer deleteTime(Integer timeID) {
-        if(splitRepository.findById(timeID).isPresent()) {
-            splitRepository.deleteById(timeID);
+        if(swimTimeRepository.findById(timeID).isPresent()) {
+
+            if(splitRepository.findById(timeID).isPresent()) {
+                splitRepository.deleteById(timeID);
+            }
+
+            swimTimeRepository.deleteById(timeID);
         }
 
-        swimTimeRepository.deleteById(timeID);
-
-        return timeID;
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Time not found.");
     }
 
     public Integer deleteTimes(Integer swimmerID) {
-        //TODO: Data validation for deletes
+        validateSwimmer(swimmerID);
         List<Integer> timeIDs = swimTimeRepository.findBySwimmerID(swimmerID);
 
-        for(Integer timeID : timeIDs) {
-            deleteTime(timeID);
+        if(timeIDs.size() > 0) {
+            for(Integer timeID : timeIDs) {
+                deleteTime(timeID);
+            }
         }
+
         return swimmerID;
     }
 
     public Integer deleteSwimmer(Integer swimmerID) {
-        if(swimmerRepository.findById(swimmerID).isPresent()) {
+        if(validateSwimmer(swimmerID)) {
             deleteTimes(swimmerID);
             swimmerRepository.deleteById(swimmerID);
-        }
 
-        return swimmerID;
+            return swimmerID;
+        }
+        return null;
+    }
+
+    private boolean validateSwimmer(Integer swimmerID) {
+        if(swimmerRepository.findById(swimmerID).isPresent()) {
+            return true;
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Swimmer not found.");
     }
 }
